Now handling invalid EC numbers by adding them to omittedOutput.omittedECNumbers. Resolves #4
Added try-catch when looking up EC values in the gene-to-EC mapping. Could have added specific testing to look for valid strings, but theoretically valid identifiers could exist that are not in the map, so this seems more robustfor now, despite being a bit hacky. When the lookup fails, a warning is printed, and the EC number is added to the omitted outputs. This update also removes the now-outdated 'This version of AutoKEGGRec is the published version' message that runs at the start of the script.